% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cow.list}
\alias{cow.list}
\title{List of state abbreviations, cow codes, and state names.}
\format{A data frame with 243 rows and 3 columns:
\describe{
  \item{StateAbb}{State Abbreviation}
  \item{CCode}{Correlates of War Country Codes}
  \item{StateNme}{State Names}
}}
\source{
\url{http://cow.la.psu.edu}
}
\usage{
cow.list
}
\description{
A dataset containing state abbreviations, cow codes, and state names.
}
\keyword{datasets}
