% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divider.r
\name{divider}
\alias{divider}
\title{Creates a divider}
\usage{
divider(x, n)
}
\arguments{
\item{x}{<- character used in divider. Examples: "-", "--", "#", etc.}

\item{n}{<- the number of times that the character should be repeated.}
}
\description{
Prints a divider. Used to seperate output that is being printed to a file.
}
\examples{
\dontrun{
x <- "-"
n <- 5
divider(x)
- - - - -
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
