% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.r
\name{elements}
\alias{elements}
\title{Count number of elements}
\usage{
elements(x)
}
\arguments{
\item{x}{A data.frame, matrix, or vector.}
}
\description{
Count the number of elements in a data.frame, matrix, or vector
}
\examples{
\dontrun{
elements("x")
}
}
