% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espnscores.r
\name{espn.scores}
\alias{espn.scores}
\title{ESPN scores function}
\usage{
espn.scores(x)
}
\arguments{
\item{x}{One of several sports leagues (i.e. MLB, NBA, NFL, NHL, WNBA)}
}
\description{
Open ESPN's scoreboard for one of several leagues in a new browser tab
}
\examples{
espn.scores("nba")
}
