% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espnstats.r
\name{espn.stats}
\alias{espn.stats}
\title{ESPN statistics function}
\usage{
espn.stats(x)
}
\arguments{
\item{x}{One of several sports leagues (e.g., MLB, NBA, NFL, NHL)}
}
\description{
Open ESPN's statistics for one of several leagues in a new browser tab
}
\examples{
espn.stats("nba")
}
