% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmissing.r
\name{fix.missing}
\alias{fix.missing}
\title{Fix missing}
\usage{
fix.missing(x, na.value)
}
\arguments{
\item{x}{A vector of values}

\item{na.value}{A missing value code}
}
\description{
Fix missing values by turning missing codes to NA.
}
\examples{
\dontrun{
x <- c(-99, 22, 15)
x <- fix.missing(x, "-99")
}
}
