% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noheader.r
\name{no.header}
\alias{no.header}
\title{Import without header}
\usage{
no.header(x)
}
\arguments{
\item{x}{File to be imported}
}
\description{
Read in a data.frame from a file without a header
}
\examples{
\dontrun{
data <- no.header("file")
}
}
