% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percents.r
\name{percents}
\alias{percents}
\title{Percents}
\usage{
percents(x)
}
\arguments{
\item{x}{Vector of fractional numbers}
}
\description{
Turn proportions into percentages
}
\examples{
\dontrun{
v <- c(0.061, 0.048, 0.10)
percents(v)
}
}
