% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit.r
\name{reddit}
\alias{reddit}
\title{Reddit search function}
\usage{
reddit(find)
}
\arguments{
\item{find}{The words (or words) you want to find, wrapped with quote (" ") marks.}
}
\description{
Enter a Reddit search in R and have the results opened in a new browser tab
}
\examples{
reddit("toronto maple leafs")
}
