% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.r
\name{rescale01}
\alias{rescale01}
\title{Rescale vector between 0 and 1}
\usage{
rescale01(x, ...)
}
\arguments{
\item{x}{Vector}

\item{...}{Pass through additional arguments, such as 'na.rm = T'}
}
\description{
Rescale a vector so its elements fall between 0 and 1
}
\examples{
\dontrun{
vector <- c(3, 4, 5, 2, 1, 0.99)
vector <- rescale01(vector, na.rm = T)
}
}
