% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siscores.R
\name{si.scores}
\alias{si.scores}
\title{SI scores function}
\usage{
si.scores(x)
}
\arguments{
\item{x}{One of several sports leagues (e.g., MLB, NBA, NFL, NHL)}
}
\description{
Open SI's scoreboard for one of several leagues in a new browser tab
}
\examples{
si.scores("nba")
}
