% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sistats.r
\name{si.stats}
\alias{si.stats}
\title{SI statistics function}
\usage{
si.stats(x)
}
\arguments{
\item{x}{One of several sports leagues (e.g., MLB, NBA, NFL, NHL)}
}
\description{
Open SI's statistics for one of several leagues in a new browser tab
}
\examples{
si.stats("nba")
}
