% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolfram.r
\name{wolfram}
\alias{wolfram}
\title{WolframAlpha search function}
\usage{
wolfram(find)
}
\arguments{
\item{find}{The words (or words) you want to find, wrapped with quote (" ") marks.}
}
\description{
Enter a WolframAlpha search in R and have the results opened in a new browser tab
}
\examples{
\dontrun{
wolfram("toronto maple leafs")
}
}
